# Loop vectorization for major convolutions of vcops