--- a/src/derivative_operators/vector_calculus_operators.jl
+++ b/src/derivative_operators/vector_calculus_operators.jl
@@ -7,7 +7,7 @@ struct GradientOperator{T,N,O}
 end
 
 function Gradient(approximation_order :: Int, dx::Union{NTuple{N,AbstractVector{T}},NTuple{N,T}},
-        len::NTuple{N,Int}, coeff_func=nothing) where {T<:Real, N}
+        len::NTuple{N,Int}; coeff_func=1) where {T<:Real, N}
     
     ops = permutedims([CenteredDifference{n}(1, approximation_order, dx[n], len[n], coeff_func) for n in 1:N])
 
@@ -23,7 +23,7 @@ struct CurlOperator{T,O}
 end 
 
 function Curl(approximation_order :: Int, dx::Union{NTuple{3,AbstractVector{T}},NTuple{3,T}},
-    len::NTuple{3,Int}, coeff_func = nothing ) where {T<:Real}
+    len::NTuple{3,Int}; coeff_func=1 ) where {T<:Real}
 
     @assert len[1] == len[2] && len[2] == len[3] "All dimensions must have equal no. of grid points"
     ops = permutedims([CenteredDifference{n}(1, approximation_order, dx[n], len[n], coeff_func) for n in 1:3])
@@ -41,7 +41,7 @@ struct DivergenceOperator{T,N,O}
 end
 
 function Divergence(approximation_order :: Int, dx::Union{NTuple{N,AbstractVector{T}},NTuple{N,T}},
-    len::NTuple{N,Int}, coeff_func=nothing) where {T<:Real, N}
+    len::NTuple{N,Int}; coeff_func=1) where {T<:Real, N}
 
     ops = permutedims([CenteredDifference{n}(1, approximation_order, dx[n], len[n], coeff_func) for n in 1:N])