Some major enhancements via using @turbo

**Benchmarks** :
- Earlier
```
Divergence
@btime u = A*u0
  31.761 μs (15 allocations: 3.52 KiB)

Gradient
@btime u = A*u0
  30.562 μs (16 allocations: 8.95 KiB)

Curl
@btime u = A*u0
  42.645 μs (15 allocations: 8.94 KiB)
```

- Now
```
Divergence
@btime u = A*u0
  8.255 μs (15 allocations: 3.52 KiB)

Gradient
@btime u = A*u0
  8.933 μs (16 allocations: 8.95 KiB)

Curl
@btime u = A*u0
  11.426 μs (15 allocations: 8.94 KiB)
```
Some other operations have also been optimized (mod removal, @turbo in `dot` , `cross`)