@mjsheikh 2021-07-24
**Update** : @turbo would work for uniform grids, since LoopVectorization doesn't allow a SMatrix kind of access in nested loops so that has been skipped at some places in new commit.

@ChrisRackauckas I think if we want to use `LoopVectorization` intensively in future for this repo, then it would be a better idea to use `HybridArrays` rather than `Array{SArray}`. For e.g. stencils for non-uniform grid `DerivateOperator`should be stored as `HybridArrays`.
Perhaps a separate PR should be done for that?